+++ diskover.py
@@ -6,7 +6,7 @@ your file metadata into Elasticsearch.
 See README.md or https://github.com/shirosaidev/diskover
 for more information.
 
-Copyright (C) Chris Park 2017-2018
+Copyright (C) Chris Park 2017-2019
 diskover is released under the Apache 2.0 license. See
 LICENSE for the full license text.
 """
@@ -38,7 +38,7 @@ import sys
 import json
 
 
-version = '1.5.0-rc29'
+version = '1.5.0.5'
 __version__ = version
 
 IS_PY3 = sys.version_info >= (3, 0)
@@ -189,7 +189,10 @@ def load_config():
             configsettings['ownersgroups_keepdomain'] = "false"
         try:
             t = config.get('autotag', 'files')
-            atf = json.loads(t)
+            if os.path.isfile("%s/%s" % (os.getcwd(),t)):
+                atf = json.loads(open("%s/%s" % (os.getcwd(),t)).read())
+            else:
+                atf = json.loads(t)
             configsettings['autotag_files'] = atf
         except ValueError as e:
             raise ValueError("Error in config autotag files: %s" % e)
@@ -197,7 +200,10 @@ def load_config():
             configsettings['autotag_files'] = []
         try:
             t = config.get('autotag', 'dirs')
-            atd = json.loads(t)
+            if os.path.isfile("%s/%s" % (os.getcwd(),t)):
+                atd = json.loads(open("%s/%s" % (os.getcwd(),t)).read())
+            else:
+                atd = json.loads(t)
             configsettings['autotag_dirs'] = atd
         except ValueError as e:
             raise ValueError("Error in config autotag dirs: %s" % e)
@@ -213,7 +219,10 @@ def load_config():
             configsettings['costpergb_base'] = 2
         try:
             s = config.get('storagecost', 'paths')
-            scp = json.loads(s)
+            if os.path.isfile("%s/%s" % (os.getcwd(),s)):
+                scp = json.loads(open("%s/%s" % (os.getcwd(),s)).read())
+            else:
+                scp = json.loads(s)            
             configsettings['costpergb_paths'] = scp
         except ValueError as e:
             raise ValueError("Error in config storagecost paths: %s" % e)
@@ -221,7 +230,10 @@ def load_config():
             configsettings['costpergb_paths'] = []
         try:
             s = config.get('storagecost', 'times')
-            sct = json.loads(s)
+            if os.path.isfile("%s/%s" % (os.getcwd(),s)):
+                sct = json.loads(open("%s/%s" % (os.getcwd(),s)).read())
+            else:
+                sct = json.loads(s)
             configsettings['costpergb_times'] = sct
         except ValueError as e:
             raise ValueError("Error in config storagecost times: %s" % e)
@@ -426,17 +438,21 @@ def load_config():
         except ConfigParser.NoOptionError:
             configsettings['gource_maxfilelag'] = 5
         try:
-            configsettings['qumulo_host'] = config.get('qumulo', 'cluster')
+            configsettings['api_url'] = config.get('crawlapi', 'url')
         except ConfigParser.NoOptionError:
-            configsettings['qumulo_host'] = ""
+            configsettings['api_url'] = ""
         try:
-            configsettings['qumulo_api_user'] = config.get('qumulo', 'api_user')
+            configsettings['api_user'] = config.get('crawlapi', 'user')
         except ConfigParser.NoOptionError:
-            configsettings['qumulo_api_user'] = ""
+            configsettings['api_user'] = ""
         try:
-            configsettings['qumulo_api_password'] = config.get('qumulo', 'api_password')
+            configsettings['api_password'] = config.get('crawlapi', 'password')
         except ConfigParser.NoOptionError:
-            configsettings['qumulo_api_password'] = ""
+            configsettings['api_password'] = ""
+        try:
+            configsettings['api_pagesize'] = config.get('crawlapi', 'pagesize')
+        except ConfigParser.NoOptionError:
+            configsettings['api_pagesize'] = ""
     except ConfigParser.NoSectionError as e:
         print('Missing section from diskover.cfg, check diskover.cfg.sample and copy over, exiting. (%s)' % e)
         sys.exit(1)
@@ -485,6 +501,19 @@ def list_plugins():
         print(plugin_info["name"])
 
 
+def user_prompt(question: str) -> bool:
+    """ Prompt the yes/no-*question* to the user. """
+    from distutils.util import strtobool
+
+    while True:
+        user_input = input(question + " [y/n]: ").lower()
+        try:
+            result = strtobool(user_input)
+            return result
+        except ValueError:
+            print("Please use y/n or yes/no.\n")
+
+
 def index_create(indexname):
     """This is the es index create function.
     It checks for existing index and deletes if
@@ -505,13 +534,19 @@ def index_create(indexname):
             return
         # delete existing index
         else:
-            logger.warning('es index exists, deleting')
-            es.indices.delete(index=indexname, ignore=[400, 404])
+            if cliargs['forcedropexisting']:
+                logger.warning('es index exists, deleting')
+                es.indices.delete(index=indexname, ignore=[400, 404])
+            else:
+                if user_prompt("Drop existing index?"):
+                    logger.warning('es index exists, deleting')
+                    es.indices.delete(index=indexname, ignore=[400, 404])
+                else:
+                    logger.info("Cannot continue with index. Exiting.")
+                    sys.exit(1)
+
     # set up es index mappings and create new index
-    if cliargs['qumulo']:
-        from diskover_qumulo import get_qumulo_mappings
-        mappings = get_qumulo_mappings(config)
-    elif cliargs['s3']:
+    if cliargs['s3']:
         from diskover_s3 import get_s3_mappings
         mappings = get_s3_mappings(config)
     else:
@@ -899,7 +934,7 @@ def index_get_docs(cliargs, logger, doctype='directory', copytags=False, hotdirs
     if pathid is True, will return dict with path and their id.
     """
 
-    data = _index_get_docs_data(index, cliargs, logger, doctype=doctype, path=path, 
+    data = _index_get_docs_data(index, cliargs, logger, doctype=doctype, path=path,
                                 maxdepth=maxdepth, sort=sort)
 
     # refresh index
@@ -1050,7 +1085,7 @@ def add_diskspace(index, logger, path):
         total_bytes = ctypes.c_ulonglong(0)
         free_bytes = ctypes.c_ulonglong(0)
         available_bytes = ctypes.c_ulonglong(0)
-        ctypes.windll.kernel32.GetDiskFreeSpaceExW(ctypes.c_wchar_p(path), 
+        ctypes.windll.kernel32.GetDiskFreeSpaceExW(ctypes.c_wchar_p(path),
             ctypes.pointer(available_bytes),
             ctypes.pointer(total_bytes),
             ctypes.pointer(free_bytes))
@@ -1155,10 +1190,11 @@ def escape_chars(text):
     """This is the escape special characters function.
     It returns escaped path strings for es queries.
     """
-    # escape any backslace characters
+    # escape any backslash chars
     text = text.replace('\\', '\\\\')
     # escape any characters in chr_dict
-    chr_dict = {'/': '\\/', '(': '\\(', ')': '\\)', '[': '\\[', ']': '\\]', '$': '\\$',
+    chr_dict = {'\n': '\\n', '\t': '\\t',
+                '/': '\\/', '(': '\\(', ')': '\\)', '[': '\\[', ']': '\\]', '$': '\\$',
                 ' ': '\\ ', '&': '\\&', '<': '\\<', '>': '\\>', '+': '\\+', '-': '\\-',
                 '|': '\\|', '!': '\\!', '{': '\\{', '}': '\\}', '^': '\\^', '~': '\\~',
                 '?': '\\?', ':': '\\:', '=': '\\=', '\'': '\\\'', '"': '\\"', '@': '\\@',
@@ -1240,6 +1276,8 @@ def parse_cli_args(indexname):
                         help="Reindex directory (non-recursive), data is added to existing index")
     parser.add_argument("-R", "--reindexrecurs", action="store_true",
                         help="Reindex directory and all subdirs (recursive), data is added to existing index")
+    parser.add_argument("-F", "--forcedropexisting", action="store_true",
+                        help="Silently drop an existing index (if present)")
     parser.add_argument("-D", "--finddupes", action="store_true",
                         help="Find duplicate files in existing index and update their dupe_md5 field")
     parser.add_argument("-C", "--copytags", metavar='INDEX2',
@@ -1259,8 +1297,10 @@ def parse_cli_args(indexname):
                         help="Replace path, example: --replacepath Z:\\ /mnt/share/")
     parser.add_argument("--crawlbot", action="store_true",
                         help="Starts up crawl bot continuous scanner to scan for dir changes in index")
-    parser.add_argument("--qumulo", action="store_true",
-                        help="Qumulo storage type, use Qumulo api instead of scandir")
+    parser.add_argument("--crawlapi", action="store_true",
+                        help="Use storage Restful API instead of scandir")
+    parser.add_argument("--storagent", metavar='HOST', nargs='+',
+                        help="Use diskover Storage Agent instead of scandir")
     parser.add_argument("--s3", metavar='FILE', nargs='+',
                         help="Import AWS S3 inventory csv file(s) (gzipped) to diskover index")
     parser.add_argument("--dircalcsonly", action="store_true",
@@ -1438,18 +1478,17 @@ def calc_dir_sizes(cliargs, logger, path=None):
                 except (ZeroDivisionError, ValueError):
                     bar.update(0)
 
-            del dirlist[:]
             # use es scroll api
             res = es.scroll(scroll_id=res['_scroll_id'], scroll='1m',
                             request_timeout=config['es_timeout'])
-        
+
         # enqueue dir calc job for any remaining in dirlist
         if len(dirlist) > 0:
             q_calc.enqueue(calc_dir_size, args=(dirlist, cliargs,), result_ttl=config['redis_ttl'])
             jobcount += 1
 
         logger.info('Found %s directory docs' % str(dircount))
-        
+
         # set up progress bar with time remaining
         if bar:
             bar.finish()
@@ -1469,7 +1508,7 @@ def calc_dir_sizes(cliargs, logger, path=None):
 
         if bar:
             bar.finish()
-        
+
         elapsed = get_time(time.time() - starttime)
         logger.info('Finished calculating %s directory sizes in %s' % (dircount, elapsed))
 
@@ -1478,28 +1517,66 @@ def calc_dir_sizes(cliargs, logger, path=None):
         sys.exit(0)
 
 
-def scandirwalk_worker(threadn):
+def scandirwalk_worker(threadn, cliargs, logger):
     dirs = []
     nondirs = []
+    # check if we are using storage agent and make connection
+    if cliargs['storagent']:
+        stor_agent = True
+        hostlist = cliargs['storagent']
+        stor_agent_conn = diskover_agent.AgentConnection(hosts=hostlist)
+        stor_agent_conn.connect()
+        if cliargs['debug'] or cliargs['verbose']:
+            logger.info("[thread-%s] Connected to Storage Agent host: %s" % (threadn, stor_agent_conn.conn_host()))
+    else:
+        stor_agent = False
     while True:
         path = q_paths.get()
         try:
             q_paths_in_progress.put(path)
             if cliargs['debug'] or cliargs['verbose']:
                 logger.info("[thread-%s] scandirwalk_worker: %s" % (threadn, path))
-            for entry in scandir(path):
-                if entry.is_dir(follow_symlinks=False) and not dir_excluded(entry.path, config, cliargs):
-                    dirs.append(entry.name)
+            if cliargs['crawlapi']:
+                root, api_dirs, api_nondirs = api_listdir(path, api_ses)
+                path = root
+                for d in api_dirs:
+                    if not dir_excluded(d[0], config, cliargs):
+                        dirs.append(d)
                 if not cliargs['dirsonly']:
-                    if entry.is_file(follow_symlinks=False):
+                    for f in api_nondirs:
+                        nondirs.append(f)
+                del api_dirs[:]
+                del api_nondirs[:]
+            elif stor_agent:
+                # grab dir list from storage agent server
+                dir_list = stor_agent_conn.listdir(path)
+                logger.debug("[thread-%s] scandirwalk_worker: Storage Agent host response time: %s" % (threadn, stor_agent_conn.response_time()))
+                path, dirs_noexcl, nondirs = dir_list
+                for d in dirs_noexcl:
+                    if not dir_excluded(d, config, cliargs):
+                        dirs.append(d)
+            else:
+                item_count = 0
+                for entry in scandir(path):
+                    if entry.is_dir(follow_symlinks=False) and not dir_excluded(entry.path, config, cliargs):
+                        dirs.append(entry.name)
+                    elif entry.is_file(follow_symlinks=False) and not cliargs['dirsonly']:
                         nondirs.append(entry.name)
+                    if item_count == 100000:
+                        if cliargs['debug'] or cliargs['verbose']:
+                            logger.info("[thread-%s] scandirwalk_worker: processing directory with many files: %s" % (threadn, path))
+                    else:
+                        item_count += 1
             q_paths_results.put((path, dirs[:], nondirs[:]))
         except (OSError, IOError) as e:
             logger.warning("[thread-%s] OS/IO Exception caused by: %s" % (threadn, e))
             pass
-        except Exception as e:
-            logger.warning("[thread-%s] Exception caused by: %s" % (threadn, e))
+        except UnicodeDecodeError as e:
+            logger.warning("[thread-%s] Unicode Decode Exception caused by: %s (path: %s)" % (threadn, e, path))
             pass
+        except Exception as e:
+            logger.error("[thread-%s] Exception caused by: %s" % (threadn, e))
+            raise
         finally:
             q_paths_in_progress.get()
         del dirs[:]
@@ -1507,19 +1584,26 @@ def scandirwalk_worker(threadn):
         q_paths.task_done()
 
 
-def scandirwalk(path):
+def scandirwalk(path, cliargs, logger):
     q_paths.put(path)
     while True:
         entry = q_paths_results.get()
         root, dirs, nondirs = entry
         if cliargs['debug'] or cliargs['verbose']:
-            logger.info("scandirwalk: %s (dircount: %s, filecount: %s)" % (root, str(len(dirs)), str(len(nondirs))))
+            if cliargs['crawlapi']:
+                logger.info("apiwalk: %s (dircount: %s, filecount: %s)" % (root[0], str(len(dirs)), str(len(nondirs))))
+            else:
+                logger.info("scandirwalk: %s (dircount: %s, filecount: %s)" % (root, str(len(dirs)), str(len(nondirs))))
         # yield before recursion
         yield root, dirs, nondirs
         # recurse into subdirectories
-        for name in dirs:
-            new_path = os.path.join(root, name)
-            q_paths.put(new_path)
+        if cliargs['crawlapi']:
+            for d in dirs:
+                q_paths.put(d[0])
+        else:
+            for name in dirs:
+                new_path = os.path.join(root, name)
+                q_paths.put(new_path)
         q_paths_results.task_done()
         if q_paths_results.qsize() == 0 and q_paths.qsize() == 0:
             time.sleep(.5)
@@ -1542,7 +1626,7 @@ def treewalk(top, num_sep, level, batchsize, cliargs, logger, reindex_dict):
 
     # set up threads for tree walk
     for i in range(cliargs['walkthreads']):
-        t = Thread(target=scandirwalk_worker, args=(i,))
+        t = Thread(target=scandirwalk_worker, args=(i, cliargs, logger,))
         t.daemon = True
         t.start()
 
@@ -1557,7 +1641,7 @@ def treewalk(top, num_sep, level, batchsize, cliargs, logger, reindex_dict):
         bar = None
 
     bartimestamp = time.time()
-    for root, dirs, files in scandirwalk(top):
+    for root, dirs, files in scandirwalk(top, cliargs, logger):
         dircount += 1
         totaldirs += 1
         files_len = len(files)
@@ -1695,13 +1779,8 @@ def crawl_tree(path, cliargs, logger, reindex_dict):
 
         logger.info("Starting crawl using %s treewalk threads (maxdepth %s)" % (cliargs['walkthreads'], cliargs['maxdepth']))
 
-        # qumulo api crawl
-        if cliargs['qumulo']:
-            from diskover_qumulo import qumulo_treewalk
-            qumulo_treewalk(path, q_crawl, num_sep, level, batchsize, cliargs, logger, reindex_dict)
-        # regular crawl using scandir
-        else:
-            treewalk(path, num_sep, level, batchsize, cliargs, logger, reindex_dict)
+        # start tree walking
+        treewalk(path, num_sep, level, batchsize, cliargs, logger, reindex_dict)
 
         return starttime
 
@@ -1882,9 +1961,9 @@ def pre_crawl_tasks():
 
     # add disk space info to es index
     if not cliargs['reindex'] and not cliargs['reindexrecurs'] and not cliargs['crawlbot']:
-        if cliargs['qumulo']:
-            from diskover_qumulo import qumulo_add_diskspace
-            qumulo_add_diskspace(es, cliargs['index'], rootdir_path, qumulo_ip, qumulo_ses, logger)
+        if cliargs['crawlapi']:
+            from diskover_crawlapi import api_add_diskspace
+            api_add_diskspace(es, cliargs['index'], rootdir_path, api_ses, logger)
         else:
             add_diskspace(cliargs['index'], logger, rootdir_path)
 
@@ -1949,19 +2028,8 @@ if __name__ == "__main__":
         calc_dir_sizes(cliargs, logger)
         sys.exit(0)
 
-    # check index name for Qumulo storage
-    if cliargs['qumulo']:
-        try:
-            if cliargs['index'] == "diskover_qumulo" or \
-                    (cliargs['index'].split('_')[0] != "diskover" and
-                             cliargs['index'].split('_')[1] != "qumulo"):
-                print('Please name your index: diskover_qumulo-<string>')
-                sys.exit(1)
-        except IndexError:
-            print('Please name your index: diskover_qumulo-<string>')
-            sys.exit(1)
     # check index name for s3 storage
-    elif cliargs['s3']:
+    if cliargs['s3']:
         try:
             if cliargs['index'] == "diskover_s3" or \
                     (cliargs['index'].split('_')[0] != "diskover" and
@@ -2041,22 +2109,25 @@ if __name__ == "__main__":
         logger.info("Plugins loaded: %s", plugins_list)
 
     # check if rootdir exists
-    if cliargs['qumulo']:
-        if IS_PY3:
-            print('Python 3 not supported using --qumulo, please use Python 2.7.')
-            sys.exit(0)
+    if cliargs['crawlapi']:
         if cliargs['rootdir'] == '.' or cliargs['rootdir'] == "":
             logger.error("Rootdir path missing, use -d /rootdir, exiting")
             sys.exit(1)
-        from diskover_qumulo import qumulo_connection, qumulo_get_file_attr
-        logger.info('Connecting to Qumulo storage api... (--qumulo)')
-        qumulo_ip, qumulo_ses = qumulo_connection()
-        logger.info('Connected to Qumulo api at %s' % qumulo_ip)
-        # check using qumulo api
+        from diskover_crawlapi import api_connection, api_stat, api_listdir
+        logger.info('Connecting to file system storage api at %s... (--crawlapi)' % config['api_url'])
+        api_ses = api_connection()
+        logger.info('Connected to storage api')
+        # check using storage api
         try:
-            qumulo_get_file_attr(cliargs['rootdir'], qumulo_ip, qumulo_ses)
-        except ValueError:
-            logger.error("Rootdir path not found or not a directory, exiting")
+            api_stat(cliargs['rootdir'], api_ses)
+        except ValueError as e:
+            logger.error("Rootdir path not found or not a directory, exiting (%s)" % e)
+            sys.exit(1)
+    elif cliargs['storagent']:
+        try:
+            import diskover_agent
+        except ImportError:
+            logger.error("Missing diskover_agent.py module, exiting")
             sys.exit(1)
     elif cliargs['s3']:
         # ingest s3 inventory files